api/requirements.txt
# Framework principal
Django==5.0.6
djangorestframework==3.15.2

# Base de datos PostgreSQL
psycopg2-binary==2.9.9

# Manejo de variables de entorno
django-environ==0.11.2

# Google Cloud Storage (para buckets y URLs firmadas)
django-storages[gcloud]==1.14.4
google-cloud-storage==2.17.0

# Utilidades
gunicorn==22.0.0        # Para correr en producción

# Servir archivos estáticos
whitenoise==6.7.0